requirements.txt
boto3==1.24.16
RPi.GPIO==0.7.0
pytest==7.1.2
unittest2==1.1.0